Desktop/code/taskAPI/requirements.txt
asgiref==3.11.0
django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pyjwt==2.10.1
python-dotenv==1.2.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.5
